Update/fix Base class in models.py
Change Base to inherit only the model mixins (Base should not inherit db.Model)

If Base inherits db.Model then this is saying that Base is a db.Model/a
sqlalchemy declarative base. sqlalchemy will think that we are trying to
do single table inheritance, which I don't believe we are. We just want
a class representing the common columns across entities.

Even if we were trying to do single table inheritance, then we need to add
additional code to all child classes of Base to instruct sqlalchemy on
how to handle two entity models with the same column names.

See link for more details:
http://docs.sqlalchemy.org/en/latest/orm/extensions/declarative/inheritance.html